requirements.txt
streamlit
pandas
openai
openpyxl
pdfplumber
python-docx
